Refactor Code
Refactored AlienPathway, and Block components for readability and to simplify logic. Moved generateScore and checkForWinner components to their own file.
