+++ components/alienPathway/CheckForWinner.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+function checkForWinner(userScore: number) {
+  if (userScore == 3) {
+    alert("You Won!");
+  }
+}
+
+export default checkForWinner;
